Perform the format step before init or validate
When running a workflow with a badly formatted file, the pipeline can
take several seconds to run the init and validate steps of the Github
action before failing at the format step.

The format operation is a lot faster than init and validate, so by
moving it earlier in the pipeline, the individual contributors receive
faster feedback on their work.